backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
supabase==2.3.1
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
scipy==1.12.0
pydantic==2.6.3
openpyxl==3.1.2
python-pptx==0.6.23
pyyaml==6.0.1
httpx==0.27.0
websockets==12.0
python-multipart==0.0.9
pytest==8.0.2
pytest-asyncio==0.23.5
requests==2.31.0
aiohttp==3.9.3
boto3==1.34.52
jinja2==3.1.3
email-validator==2.1.0 